refactor: narrow LOVAC postprocess scope
