+++ .github/workflows/python-scripts.yml
@@ -8,6 +8,7 @@ on:
       - server/src/scripts/**/*.py
       - server/src/scripts/**/requirements.txt
       - analytics/dagster/scripts/owner-housing-distances/**/*.py
+      - analytics/dagster/src/owner_housing_locations/**/*.py
       - .github/workflows/python-scripts.yml
   push:
     branches:
@@ -16,6 +17,7 @@ on:
       - server/src/scripts/**/*.py
       - server/src/scripts/**/requirements.txt
       - analytics/dagster/scripts/owner-housing-distances/**/*.py
+      - analytics/dagster/src/owner_housing_locations/**/*.py
       - .github/workflows/python-scripts.yml
 
 jobs:
@@ -62,7 +64,9 @@ jobs:
         if: matrix.python-version == '3.13'
         run: |
           cd server/src/scripts/owner-housing-distances
-          pytest test_calculate_distances.py --cov=. --cov-report=xml --cov-report=term
+          pytest test_calculate_distances.py \
+            --cov="$GITHUB_WORKSPACE/analytics/dagster/src/owner_housing_locations" \
+            --cov-report=xml --cov-report=term
 
       - name: Upload coverage to Codecov
         if: matrix.python-version == '3.13'
@@ -96,22 +100,26 @@ jobs:
         run: |
           black --check --diff \
             server/src/scripts/ \
-            analytics/dagster/scripts/owner-housing-distances/
+            analytics/dagster/scripts/owner-housing-distances/ \
+            analytics/dagster/src/owner_housing_locations/
         continue-on-error: true
 
       - name: Check import sorting with isort
         run: |
           isort --check-only --diff \
             server/src/scripts/ \
-            analytics/dagster/scripts/owner-housing-distances/
+            analytics/dagster/scripts/owner-housing-distances/ \
+            analytics/dagster/src/owner_housing_locations/
         continue-on-error: true
 
       - name: Lint with flake8
         run: |
           # Stop the build if there are Python syntax errors or undefined names
           flake8 server/src/scripts/ analytics/dagster/scripts/owner-housing-distances/ \
+            analytics/dagster/src/owner_housing_locations/ \
             --count --select=E9,F63,F7,F82 --show-source --statistics
           # Exit-zero treats all errors as warnings
           flake8 server/src/scripts/ analytics/dagster/scripts/owner-housing-distances/ \
+            analytics/dagster/src/owner_housing_locations/ \
             --count --exit-zero --max-complexity=10 --max-line-length=120 --statistics
         continue-on-error: true
