+++ analytics/dagster/pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "psutil>=7.0.0",
     "pyarrow>=21.0.0",
     "tenacity>=9.0.0",
+    "tqdm>=4.67.1",
 ]
 
 [project.optional-dependencies]
