+++ analytics/dagster/src/__init__.py
@@ -1,1 +1,1 @@
-from .definitions import defs as defs
+"""Zéro Logement Vacant Dagster modules."""
