+++ analytics/dagster/src/zlv_dbt/README.md
@@ -2,7 +2,7 @@
 
 This directory was generated by `dagster-dbt project scaffold` and is **not**
 registered by `analytics/dagster/workspace.yaml` (which loads only
-`python_module: src`).
+`python_module: src.definitions`).
 
 The active dbt assets live in
 [`src/assets/production_dbt.py`](../assets/production_dbt.py) and are wired
