+++ analytics/dagster/workspace.yaml
@@ -1,2 +1,2 @@
 load_from:
-  - python_module: src
+  - python_module: src.definitions
