+++ server/src/scripts/owner-housing-distances/calculate_distances.py
@@ -1,32 +1,31 @@
 #!/usr/bin/env python3
-"""Compatibility entrypoint for the canonical Dagster location calculator."""
+"""Compatibility CLI for the importable Dagster location calculator."""
 
 from __future__ import annotations
 
-import importlib.util
 import sys
 from pathlib import Path
 
-_SCRIPT = (
-    Path(__file__).resolve().parents[4]
-    / "analytics"
-    / "dagster"
-    / "scripts"
-    / "owner-housing-distances"
-    / "calculate_distances.py"
-)
-_SPEC = importlib.util.spec_from_file_location("zlv_owner_housing_calculator", _SCRIPT)
-_MODULE = importlib.util.module_from_spec(_SPEC)
-if _SPEC.loader is None:
-    raise RuntimeError(f"Cannot load {_SCRIPT}")
-sys.modules[_SPEC.name] = _MODULE
-_SPEC.loader.exec_module(_MODULE)
+_DAGSTER_ROOT = Path(__file__).resolve().parents[4] / "analytics" / "dagster"
+if str(_DAGSTER_ROOT) not in sys.path:
+    sys.path.insert(0, str(_DAGSTER_ROOT))
 
-globals().update(
-    {name: value for name, value in vars(_MODULE).items() if not name.startswith("__")}
+from src.owner_housing_locations.calculator import (  # noqa: E402
+    DistanceCalculator,
+    LocationComputationReport,
+    LocationScope,
+    calculate_owner_housing_locations,
+    main,
 )
+
+__all__ = [
+    "DistanceCalculator",
+    "LocationComputationReport",
+    "LocationScope",
+    "calculate_owner_housing_locations",
+    "main",
+]
+
 
 if __name__ == "__main__":
-    _MODULE.main()
-else:
-    sys.modules[__name__] = _MODULE
+    main()
